+++ sdk/communication/communication-chat/src/chatClient.ts
@@ -35,21 +35,14 @@ import { ChatThreadClient } from "./chatThreadClient";
 import {
   ChatClientOptions,
   CreateChatThreadOptions,
-  GetChatThreadOptions,
   ListChatThreadsOptions,
   DeleteChatThreadOptions
 } from "./models/options";
 import {
-  mapToChatThreadSdkModel,
   mapToChatParticipantRestModel,
   mapToCreateChatThreadResultSdkModel
 } from "./models/mappers";
-import {
-  ChatThreadInfo,
-  CreateChatThreadResult,
-  ChatThread,
-  ListPageSettings
-} from "./models/models";
+import { ChatThreadItem, CreateChatThreadResult, ListPageSettings } from "./models/models";
 import { createCommunicationTokenCredentialPolicy } from "./credential/communicationTokenCredentialPolicy";
 import { ChatApiClient } from "./generated/src";
 import { CreateChatThreadRequest } from "./models/requests";
@@ -69,12 +62,12 @@ export class ChatClient {
   /**
    * Creates an instance of the ChatClient for a given resource and user.
    *
-   * @param url - The url of the Communication Services resouce.
+   * @param endpoint - The url of the Communication Services resouce.
    * @param credential - The token credential. Use AzureCommunicationTokenCredential from \@azure/communication-common to create a credential.
    * @param options - Additional client options.
    */
   constructor(
-    private readonly url: string,
+    private readonly endpoint: string,
     credential: CommunicationTokenCredential,
     options: ChatClientOptions = {}
   ) {
@@ -106,7 +99,7 @@ export class ChatClient {
     const authPolicy = createCommunicationTokenCredentialPolicy(this.tokenCredential);
     const pipeline = createPipelineFromOptions(internalPipelineOptions, authPolicy);
 
-    this.client = new ChatApiClient(this.url, pipeline);
+    this.client = new ChatApiClient(this.endpoint, pipeline);
 
     this.signalingClient = getSignalingClient(credential, logger);
   }
@@ -116,7 +109,7 @@ export class ChatClient {
    * @param threadId - Thread ID for the ChatThreadClient
    */
   public getChatThreadClient(threadId: string): ChatThreadClient {
-    return new ChatThreadClient(this.url, threadId, this.tokenCredential, this.clientOptions);
+    return new ChatThreadClient(this.endpoint, threadId, this.tokenCredential, this.clientOptions);
   }
 
   /**
@@ -132,13 +125,12 @@ export class ChatClient {
     const { span, updatedOptions } = createSpan("ChatClient-CreateChatThread", options);
 
     try {
-      // We generate an UUID if user not provides repeatabilityRequestId.
-      updatedOptions.repeatabilityRequestId =
-        updatedOptions.repeatabilityRequestId ?? generateUuid();
+      // We generate an UUID if the user does not provide an idempotencyToken value
+      updatedOptions.idempotencyToken = updatedOptions.idempotencyToken ?? generateUuid();
       const { _response, ...result } = await this.client.chat.createChatThread(
         {
           topic: request.topic,
-          participants: request.participants?.map((participant) =>
+          participants: options.participants?.map((participant) =>
             mapToChatParticipantRestModel(participant)
           )
         },
@@ -156,39 +148,10 @@ export class ChatClient {
     }
   }
 
-  /**
-   * Gets a chat thread.
-   * Returns the chat thread.
-   * @param threadId - The ID of the thread to get.
-   * @param options -  Operation options.
-   */
-  public async getChatThread(
-    threadId: string,
-    options: GetChatThreadOptions = {}
-  ): Promise<ChatThread> {
-    const { span, updatedOptions } = createSpan("ChatClient-GetChatThread", options);
-
-    try {
-      const { _response, ...result } = await this.client.chat.getChatThread(
-        threadId,
-        operationOptionsToRequestOptionsBase(updatedOptions)
-      );
-      return mapToChatThreadSdkModel(result);
-    } catch (e) {
-      span.setStatus({
-        code: CanonicalCode.UNKNOWN,
-        message: e.message
-      });
-      throw e;
-    } finally {
-      span.end();
-    }
-  }
-
   private async *listChatThreadsPage(
     continuationState: ListPageSettings,
     options: ListChatThreadsOptions = {}
-  ): AsyncIterableIterator<ChatThreadInfo[]> {
+  ): AsyncIterableIterator<ChatThreadItem[]> {
     const requestOptions = operationOptionsToRequestOptionsBase(options);
     if (!continuationState.continuationToken) {
       const currentSetResponse = await this.client.chat.listChatThreads(requestOptions);
@@ -214,7 +177,7 @@ export class ChatClient {
 
   private async *listChatThreadsAll(
     options: ListChatThreadsOptions
-  ): AsyncIterableIterator<ChatThreadInfo> {
+  ): AsyncIterableIterator<ChatThreadItem> {
     for await (const page of this.listChatThreadsPage({}, options)) {
       yield* page;
     }
@@ -226,7 +189,7 @@ export class ChatClient {
    */
   public listChatThreads(
     options: ListChatThreadsOptions = {}
-  ): PagedAsyncIterableIterator<ChatThreadInfo> {
+  ): PagedAsyncIterableIterator<ChatThreadItem> {
     const { span, updatedOptions } = createSpan("ChatClient-ListChatThreads", options);
     try {
       const iter = this.listChatThreadsAll(updatedOptions);