+++ sdk/containerregistry/container-registry/samples/typescript/src/containerRegistryClient.ts
@@ -7,15 +7,18 @@
 
 import { ContainerRegistryClient } from "@azure/container-registry";
 import { DefaultAzureCredential } from "@azure/identity";
+import * as dotenv from "dotenv";
+dotenv.config();
 
 export async function main() {
   const endpoint = process.env.ENDPOINT ?? "<endpoint>";
 
   const client = new ContainerRegistryClient(endpoint, new DefaultAzureCredential());
-
-  const attributes = await client.getAttributes("repo_name");
-
-  console.log("The value of registry:", attributes.registry);
+  const response = await client.deleteRepository("hello-world");
+  console.log(
+    `Number of artifacts deleted: ${response?.deletedRegistryArtifactDigests?.length ?? 0}`
+  );
+  console.log(`Number of tags deleted: ${response?.deletedRegistryArtifactDigests?.length ?? 0}`);
 }
 
 main().catch((err) => {