+++ sdk/identity/identity/src/credentials/defaultAzureCredential.ts
@@ -9,26 +9,29 @@ import { AzureCliCredential } from "./azureCliCredential";
 import { VisualStudioCodeCredential } from "./visualStudioCodeCredential";
 
 /**
- * Provides options to configure the default Azure credentials.
+ * Provides options to configure the {@link DefaultAzureCredential} class.
  */
 export interface DefaultAzureCredentialOptions extends TokenCredentialOptions {
   /**
-   * Optionally pass in a Tenant ID to be used as part of the credential
+   * Optionally pass in a Tenant ID to be used as part of the credential.
+   * By default it may use a generic tenant ID depending on the underlying credential.
    */
   tenantId?: string;
   /**
-   * Optionally pass in a user assigned client ID for the ManagedIdentityCredential
+   * Optionally pass in a user assigned client ID to be used by the {@link ManagedIdentityCredential}.
+   * This client ID can also be passed through to the {@link ManagedIdentityCredential} through the environment variable: AZURE_CLIENT_ID.
    */
   managedIdentityClientId?: string;
 }
 
 /**
- * Provides a default {@link ChainedTokenCredential} configuration for
- * applications that will be deployed to Azure.  The following credential
- * types will be tried, in order:
+ * Provides a default {@link ChainedTokenCredential} configuration that should work for most applications that use the Azure SDK.
+ * The following credential types will be tried, in order:
  *
  * - {@link EnvironmentCredential}
  * - {@link ManagedIdentityCredential}
+ * - {@link AzureCliCredential}
+ * - {@link VisualStudioCodeCredential}
  *
  * Consult the documentation of these credential types for more information
  * on how they attempt authentication.
@@ -37,22 +40,23 @@ export class DefaultAzureCredential extends ChainedTokenCredential {
   /**
    * Creates an instance of the DefaultAzureCredential class.
    *
-   * @param options - Options for configuring the client which makes the authentication request.
+   * @param options - Optional parameters. See {@link DefaultAzureCredentialOptions}.
    */
   constructor(tokenCredentialOptions?: DefaultAzureCredentialOptions) {
     const credentials = [];
     credentials.push(new EnvironmentCredential(tokenCredentialOptions));
 
-    // In case a user assigned ID has been provided.
+    // A client ID for the ManagedIdentityCredential
+    // can be provided either through the optional parameters or through the environment variables.
     const managedIdentityClientId =
       tokenCredentialOptions?.managedIdentityClientId || process.env.AZURE_CLIENT_ID;
 
+    // If a client ID is not provided, we will try with the system assigned ID.
     if (managedIdentityClientId) {
       credentials.push(
         new ManagedIdentityCredential(managedIdentityClientId, tokenCredentialOptions)
       );
     } else {
-      // If the user didn't provide an ID, we'll try with a system assigned ID.
       credentials.push(new ManagedIdentityCredential(tokenCredentialOptions));
     }