+++ sdk/identity/identity/src/credentials/usernamePasswordCredentialOptions.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { InteractiveCredentialOptions } from "./interactiveCredentialOptions";
+
+/**
+ * Defines options for the {@link UsernamePasswordCredential} class.
+ */
+export interface UsernamePasswordCredentialOptions extends InteractiveCredentialOptions {}