+++ sdk/keyvault/keyvault-secrets/test/public/recoverBackupRestore.spec.ts
@@ -153,7 +153,10 @@ describe("Secret client - restore secrets and recover backups", () => {
       // This test implementation of a restore poller only applies for backups that have been recently deleted.
       // Backups might not be ready to be restored in an unknown amount of time.
       // If this is useful to you, please open an issue at: https://github.com/Azure/azure-sdk-for-js/issues
-      const restorePoller = await testClient.beginRestoreSecretBackup(backup as Uint8Array);
+      const restorePoller = await testClient.beginRestoreSecretBackup(
+        backup as Uint8Array,
+        testPollerProperties
+      );
       const restoredSecretProperties = await restorePoller.pollUntilDone();
 
       assert.equal(restoredSecretProperties.name, secretName);