+++ sdk/storage/arm-storage/src/models/blobContainersMappers.ts
@@ -36,6 +36,7 @@ export {
   DeletedAccount,
   DeleteRetentionPolicy,
   Encryption,
+  EncryptionIdentity,
   EncryptionScope,
   EncryptionScopeKeyVaultProperties,
   EncryptionService,
@@ -67,6 +68,7 @@ export {
   ManagementPolicyRule,
   ManagementPolicySchema,
   ManagementPolicySnapShot,
+  ManagementPolicyVersion,
   Multichannel,
   NetworkRuleSet,
   ObjectReplicationPolicy,
@@ -96,5 +98,6 @@ export {
   TagProperty,
   TrackedResource,
   UpdateHistoryProperty,
+  UserAssignedIdentity,
   VirtualNetworkRule
 } from "../models/mappers";