+++ sdk/core/core-client/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Release History
 
-## 1.0.0 (Unreleased)
+## 1.0.1 (Unreleased)
 
+
+## 1.0.0 (2021-03-15)
+
+GA release of this package.
 
 ## 1.0.0-beta.2 (2021-03-10)