+++ sdk/keyvault/keyvault-keys/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 4.2.0-beta.5 (Unreleased)
 
+- Added local cryptography support for encryption / decryption for `A128CBCPAD`, `A192CBCPAD`, and `A256CBCPAD`.
+- For AES-CBC encryption we will now generate an IV if the user did not pass it in, making `iv` optional for those parameters.
 
 ## 4.2.0-beta.4 (2021-03-09)